break parser and session state modification into separte files
