+++ pfcpiface/messages.go
@@ -7,165 +7,13 @@ import (
 	"errors"
 	"log"
 	"net"
-	"strings"
-	"sync"
 	"time"
 
 	"github.com/wmnsk/go-pfcp/ie"
 	"github.com/wmnsk/go-pfcp/message"
 )
 
-// PktBufSz : buffer size for incoming pkt
-const (
-	PktBufSz    = 1500
-	PFCPPort    = "8805"
-	MaxItems    = 10
-	Timeout     = 1000 * time.Millisecond
-	readTimeout = 25 * time.Second
-)
-
-type sessRecord struct {
-	pdrs []pdr
-	fars []far
-}
-
-var sessions map[uint64]sessRecord
-
-type sequenceNumber struct {
-	seq uint32
-	mux sync.Mutex
-}
-
-var seqNum sequenceNumber
-
-func pfcpifaceMainLoop(upf *upf, accessIP, coreIP, sourceIP, smfName string) {
-	log.Println("pfcpifaceMainLoop@" + upf.fqdnHost + " says hello!!!")
-
-	cpConnectionStatus := make(chan bool)
-
-	// Verify IP + Port binding
-	laddr, err := net.ResolveUDPAddr("udp", sourceIP+":"+PFCPPort)
-	if err != nil {
-		log.Fatalln("Unable to resolve udp addr!", err)
-		return
-	}
-
-	// Listen on the port
-	conn, err := net.ListenUDP("udp", laddr)
-	if err != nil {
-		log.Fatalln("Unable to bind to listening port!", err)
-		return
-	}
-
-	// flag to check SMF/SPGW-C is connected
-	//cpConnected is true if upf received request from control plane or
-	// if upf receives +ve response for upf initiated setup request
-	cpConnected := false
-
-	// cleanup the pipeline
-	cleanupSessions := func() {
-		if cpConnected {
-			sendDeleteAllSessionsMsgtoUPF(upf)
-			cpConnected = false
-		}
-	}
-	//initiate connection if smf address available
-	log.Println("calling manageSmfConnection smf service name ", smfName)
-	manageConnection := false
-	if smfName != "" {
-		manageConnection = true
-		go manageSmfConnection(sourceIP, accessIP, smfName, conn, cpConnectionStatus)
-	}
-
-	// Initialize pkt buf
-	buf := make([]byte, PktBufSz)
-	// Initialize pkt header
-
-	// Initialize sessions map
-	sessions = make(map[uint64]sessRecord)
-
-	for {
-		err := conn.SetReadDeadline(time.Now().Add(readTimeout))
-		if err != nil {
-			log.Fatalln("Unable to set deadline for read:", err)
-		}
-		// blocking read
-		n, addr, err := conn.ReadFrom(buf)
-		if err != nil {
-			if err, ok := err.(net.Error); ok && err.Timeout() {
-				// do nothing for the time being
-				log.Println(err)
-				cpConnected = false
-				if manageConnection {
-					cpConnectionStatus <- cpConnected
-				}
-				cleanupSessions()
-				continue
-			}
-			log.Fatalln("Read error:", err)
-		}
-
-		// use wmnsk lib to parse the pfcp message
-		msg, err := message.Parse(buf[:n])
-		if err != nil {
-			log.Println("Ignoring undecodable message: ", buf[:n], " error: ", err)
-			continue
-		}
-
-		// if sourceIP is not set, fetch it from the msg header
-		if sourceIP == "0.0.0.0" {
-			addrString := strings.Split(addr.String(), ":")
-			sourceIP = getLocalIP(addrString[0]).String()
-			log.Println("Source IP address is now: ", sourceIP)
-		}
-
-		//log.Println("Message: ", msg)
-
-		// handle message
-		var outgoingMessage []byte
-		switch msg.MessageType() {
-		case message.MsgTypeAssociationSetupRequest:
-			outgoingMessage = handleAssociationSetupRequest(msg, addr, sourceIP, accessIP, coreIP)
-			if outgoingMessage != nil {
-				cpConnected = true
-				if manageConnection {
-					// if we initiated connection, inform go routine
-					cpConnectionStatus <- cpConnected
-				}
-			}
-		case message.MsgTypeAssociationSetupResponse:
-			cpConnected = handleAssociationSetupResponse(msg, addr, sourceIP, accessIP)
-			if manageConnection {
-				// pass on information to go routine that result of association response
-				cpConnectionStatus <- cpConnected
-			}
-		case message.MsgTypeSessionEstablishmentRequest:
-			outgoingMessage = handleSessionEstablishmentRequest(upf, msg, addr, sourceIP)
-		case message.MsgTypeSessionModificationRequest:
-			outgoingMessage = handleSessionModificationRequest(upf, msg, addr, sourceIP)
-		case message.MsgTypeHeartbeatRequest:
-			outgoingMessage = handleHeartbeatRequest(msg, addr)
-		case message.MsgTypeSessionDeletionRequest:
-			outgoingMessage = handleSessionDeletionRequest(upf, msg, addr, sourceIP)
-		case message.MsgTypeAssociationReleaseRequest:
-			outgoingMessage = handleAssociationReleaseRequest(msg, addr, sourceIP, accessIP)
-			cleanupSessions()
-		default:
-			log.Println("Message type: ", msg.MessageTypeName(), " is currently not supported")
-			continue
-		}
-
-		// send the response out
-		if outgoingMessage != nil {
-			if _, err := conn.WriteTo(outgoingMessage, addr); err != nil {
-				log.Fatalln("Unable to transmit association setup response", err)
-			}
-		}
-
-	}
-}
-
-func handleAssociationSetupRequest(msg message.Message, addr net.Addr, sourceIP string, accessIP string, coreIP string) []byte {
+func (pc *PFCPConn) handleAssociationSetupRequest(msg message.Message, addr net.Addr, sourceIP string, accessIP string, coreIP string) []byte {
 	asreq, ok := msg.(*message.AssociationSetupRequest)
 	if !ok {
 		log.Println("Got an unexpected message: ", msg.MessageTypeName(), " from: ", addr)
@@ -188,7 +36,7 @@ func handleAssociationSetupRequest(msg message.Message, addr net.Addr, sourceIP 
 		// 0x41 = Spare (0) | Assoc Src Inst (1) | Assoc Net Inst (0) | Tied Range (000) | IPV6 (0) | IPV4 (1)
 		//      = 01000001
 		ie.NewUserPlaneIPResourceInformation(0x41, 0, accessIP, "", "", ie.SrcInterfaceAccess),
-		//ie.NewUserPlaneIPResourceInformation(0x41, 0, coreIP, "", "", ie.SrcInterfaceCore),
+		// ie.NewUserPlaneIPResourceInformation(0x41, 0, coreIP, "", "", ie.SrcInterfaceCore),
 	).Marshal() /* userplane ip resource info */
 	if err != nil {
 		log.Fatalln("Unable to create association setup response", err)
@@ -251,7 +99,7 @@ func handleAssociationReleaseRequest(msg message.Message, addr net.Addr, sourceI
 	return arres
 }
 
-func handleSessionEstablishmentRequest(upf *upf, msg message.Message, addr net.Addr, sourceIP string) []byte {
+func (pc *PFCPConn) handleSessionEstablishmentRequest(upf *upf, msg message.Message, addr net.Addr, sourceIP string) []byte {
 	sereq, ok := msg.(*message.SessionEstablishmentRequest)
 	if !ok {
 		log.Println("Got an unexpected message: ", msg.MessageTypeName(), " from: ", addr)
@@ -273,7 +121,6 @@ func handleSessionEstablishmentRequest(upf *upf, msg message.Message, addr net.A
 	fars := make([]far, 0, MaxItems)
 
 	sendError := func(err error) []byte {
-
 		log.Println(err)
 		// Build response message
 		seres, err := message.NewSessionEstablishmentResponse(0, /* MO?? <-- what's this */
@@ -284,7 +131,6 @@ func handleSessionEstablishmentRequest(upf *upf, msg message.Message, addr net.A
 			ie.NewNodeID(sourceIP, "", ""), /* node id (IPv4) */
 			ie.NewCause(ie.CauseRequestRejected),
 		).Marshal()
-
 		if err != nil {
 			log.Fatalln("Unable to create session establishment response", err)
 		}
@@ -312,11 +158,11 @@ func handleSessionEstablishmentRequest(upf *upf, msg message.Message, addr net.A
 	upf.sendMsgToUPF("add", pdrs, fars)
 
 	// Adding current session details to the hash map
-	sessItem := sessRecord{
+	sessItem := PFCPSession{
 		pdrs: pdrs,
 		fars: fars,
 	}
-	sessions[fseid.SEID] = sessItem
+	pc.mgr.sessions[fseid.SEID] = sessItem
 
 	// Build response message
 	seres, err := message.NewSessionEstablishmentResponse(0, /* MO?? <-- what's this */
@@ -328,7 +174,6 @@ func handleSessionEstablishmentRequest(upf *upf, msg message.Message, addr net.A
 		ie.NewCause(ie.CauseRequestAccepted), /* accept it blindly for the time being */
 		ie.NewFSEID(peerSEID(fseid.SEID), net.ParseIP(sourceIP), nil, nil),
 	).Marshal()
-
 	if err != nil {
 		log.Fatalln("Unable to create session establishment response", err)
 	}
@@ -338,7 +183,7 @@ func handleSessionEstablishmentRequest(upf *upf, msg message.Message, addr net.A
 	return seres
 }
 
-func handleSessionModificationRequest(upf *upf, msg message.Message, addr net.Addr, sourceIP string) []byte {
+func (pc *PFCPConn) handleSessionModificationRequest(upf *upf, msg message.Message, addr net.Addr, sourceIP string) []byte {
 	smreq, ok := msg.(*message.SessionModificationRequest)
 	if !ok {
 		log.Println("Got an unexpected message: ", msg.MessageTypeName(), " from: ", addr)
@@ -355,7 +200,6 @@ func handleSessionModificationRequest(upf *upf, msg message.Message, addr net.Ad
 	fars := make([]far, 0, MaxItems)
 
 	sendError := func(err error) []byte {
-
 		log.Println(err)
 		smres, err := message.NewSessionModificationResponse(0, /* MO?? <-- what's this */
 			0,                                    /* FO <-- what's this? */
@@ -365,7 +209,6 @@ func handleSessionModificationRequest(upf *upf, msg message.Message, addr net.Ad
 			ie.NewCause(ie.CauseRequestAccepted), /* accept it blindly for the time being */
 			ie.NewFSEID(peerSEID(smreq.SEID()), net.ParseIP(sourceIP), nil, nil),
 		).Marshal()
-
 		if err != nil {
 			log.Fatalln("Unable to create session establishment response", err)
 		}
@@ -401,7 +244,6 @@ func handleSessionModificationRequest(upf *upf, msg message.Message, addr net.Ad
 		ie.NewCause(ie.CauseRequestAccepted), /* accept it blindly for the time being */
 		ie.NewFSEID(peerSEID(smreq.SEID()), net.ParseIP(sourceIP), nil, nil),
 	).Marshal()
-
 	if err != nil {
 		log.Fatalln("Unable to create session modification response", err)
 	}
@@ -424,7 +266,6 @@ func handleHeartbeatRequest(msg message.Message, addr net.Addr) []byte {
 	hbres, err := message.NewHeartbeatResponse(hbreq.SequenceNumber,
 		ie.NewRecoveryTimeStamp(time.Now()), /* ts */
 	).Marshal()
-
 	if err != nil {
 		log.Fatalln("Unable to create heartbeat response", err)
 	}
@@ -434,7 +275,7 @@ func handleHeartbeatRequest(msg message.Message, addr net.Addr) []byte {
 	return hbres
 }
 
-func handleSessionDeletionRequest(upf *upf, msg message.Message, addr net.Addr, sourceIP string) []byte {
+func (pc *PFCPConn) handleSessionDeletionRequest(upf *upf, msg message.Message, addr net.Addr, sourceIP string) []byte {
 	sdreq, ok := msg.(*message.SessionDeletionRequest)
 	if !ok {
 		log.Println("Got an unexpected message: ", msg.MessageTypeName(), " from: ", addr)
@@ -444,12 +285,12 @@ func handleSessionDeletionRequest(upf *upf, msg message.Message, addr net.Addr, 
 	log.Println("Got a session deletion request from: ", addr)
 
 	/* retrieve sessionRecord */
-	sessItem := sessions[mySEID(sdreq.SEID())]
+	sessItem := pc.mgr.sessions[mySEID(sdreq.SEID())]
 
 	upf.sendMsgToUPF("del", sessItem.pdrs, sessItem.fars)
 
 	/* delete sessionRecord */
-	delete(sessions, mySEID(sdreq.SEID()))
+	delete(pc.mgr.sessions, mySEID(sdreq.SEID()))
 
 	// Build response message
 	smres, err := message.NewSessionDeletionResponse(0, /* MO?? <-- what's this */
@@ -459,7 +300,6 @@ func handleSessionDeletionRequest(upf *upf, msg message.Message, addr net.Addr, 
 		0,                                    /* priority */
 		ie.NewCause(ie.CauseRequestAccepted), /* accept it blindly for the time being */
 	).Marshal()
-
 	if err != nil {
 		log.Fatalln("Unable to create session deletion response", err)
 	}
@@ -469,14 +309,7 @@ func handleSessionDeletionRequest(upf *upf, msg message.Message, addr net.Addr, 
 	return smres
 }
 
-func getSeqNum() uint32 {
-	seqNum.mux.Lock()
-	defer seqNum.mux.Unlock()
-	seqNum.seq++
-	return seqNum.seq
-}
-
-func manageSmfConnection(n4LocalIP string, n3ip string, n4Dst string, conn *net.UDPConn, cpConnectionStatus chan bool) {
+func (pc *PFCPConn) manageSmfConnection(n4LocalIP string, n3ip string, n4Dst string, conn *net.UDPConn, cpConnectionStatus chan bool) {
 	msg := false
 	cpConnected := false
 
@@ -486,13 +319,13 @@ func manageSmfConnection(n4LocalIP string, n3ip string, n4Dst string, conn *net.
 		log.Println("SPGWC/SMF address IP inside manageSmfConnection ", n4DstIP.String())
 		// initiate request if we have control plane address available
 		if n4DstIP.String() != "0.0.0.0" {
-			generateAssociationRequest(n4LocalIP, n3ip, n4DstIP.String(), conn)
+			pc.generateAssociationRequest(n4LocalIP, n3ip, n4DstIP.String(), conn)
 		}
 		// no worry. Looks like control plane is still not up
 	}
 	updateSmfStatus := func(msg bool) {
 		log.Println("cpConnected : ", cpConnected, "msg ", msg)
-		//events from main Loop
+		// events from main Loop
 		if cpConnected && !msg {
 			log.Println("CP disconnected ")
 			cpConnected = false
@@ -511,7 +344,7 @@ func manageSmfConnection(n4LocalIP string, n3ip string, n4Dst string, conn *net.
 	for {
 		select {
 		case msg = <-cpConnectionStatus:
-			//events from main Loop
+			// events from main Loop
 			updateSmfStatus(msg)
 			if cpConnected {
 				pfcpResponseTicker.Stop()
@@ -529,12 +362,11 @@ func manageSmfConnection(n4LocalIP string, n3ip string, n4Dst string, conn *net.
 	}
 }
 
-func generateAssociationRequest(n4LocalIP string, n3ip string, n4DstIp string, conn *net.UDPConn) {
-
-	seq_num := getSeqNum()
-	log.Println("n4DstIp ", n4DstIp)
+func (pc *PFCPConn) generateAssociationRequest(n4LocalIP string, n3ip string, n4DstIP string, conn *net.UDPConn) {
+	seq := pc.getSeqNum()
+	log.Println("n4DstIp ", n4DstIP)
 	// Build request message
-	asreq, err := message.NewAssociationSetupRequest(seq_num, ie.NewRecoveryTimeStamp(time.Now()),
+	asreq, err := message.NewAssociationSetupRequest(seq, ie.NewRecoveryTimeStamp(time.Now()),
 		ie.NewNodeID(n4LocalIP, "", ""), /* node id (IPv4) */
 		// 0x41 = Spare (0) | Assoc Src Inst (1) | Assoc Net Inst (0) | Tied Range (000) | IPV6 (0) | IPV4 (1)
 		//      = 01000001
@@ -544,7 +376,7 @@ func generateAssociationRequest(n4LocalIP string, n3ip string, n4DstIp string, c
 		log.Fatalln("Unable to create association setup response", err)
 	}
 
-	smfAddr, err := net.ResolveUDPAddr("udp", n4DstIp+":"+PFCPPort)
+	smfAddr, err := net.ResolveUDPAddr("udp", n4DstIP+":"+PFCPPort)
 	if err != nil {
 		log.Fatalln("Unable to resolve udp addr!", err)
 		return
